MentorLy/Python_APIs/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.41
botocore==1.38.41
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
google-auth==2.40.3
google-genai==1.21.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.18
plum-dispatch==1.7.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
s3transfer==0.13.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
spire-doc==13.4.6
starlette==0.46.2
tenacity==8.5.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
